ci: locate topbar buttons and configs without UI copy
